Added a checkbox in the compare webpage to filter out scenarios that do not have a reference screenshot.  I have also added a counter that shows how many scenarios are currently displayed.
There is a caveat in the solution I have implemented. In order to find
out if a local reference image is present or not I have to construct as
many image objects as scenarios present. In addition, this information
requires waiting for an asynchronous request to complete, so the
information is stale until the request completes. Maybe a better
solution is to include information about the presence of the reference
image in config.js which is most probably constructed having direct
access to the file-system.